requirements.txt
chardet==4.0.0
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
dpkt==1.9.7.2
Flask==2.0.2
Flask-Excel==0.0.7
itsdangerous==2.0.1
Jinja2==3.0.3
lml==0.1.0
MarkupSafe==2.0.1
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.1.0
pyexcel==0.6.7
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pymongo==4.0.1
texttable==1.6.4
Werkzeug==2.0.3
